cmd/pyproject.toml
[project]
name = "cmd"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Oliver Baumann", email = "oliver.baumann@uni-bayreuth.de" }
]
dependencies = [
    "requests>=2.32.3",
    "typer>=0.12.3",
    "wasabi>=1.1.3",
    "python-dotenv>=1.0.1",
    "numpy==1.26.2",
    "pandas>=2.2.2",
    "tqdm>=4.66.5",
    "scikit-learn>=1.5.1",
    "matplotlib>=3.9.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=24.4.2",
    "isort>=5.13.2",
    "pyflakes>=3.2.0",
    "autopep8>=2.3.1",
    "scikit-build>=0.18.0",
]
virtual = true

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/cmd"]
